pyproject.toml
[tool.poetry]
name = "grpc-accesslog"
version = "0.2.1"
description = "gRPC Access Log"
authors = ["Michael Morgan <git@morgan83.com>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/villainy/grpc-accesslog"
repository = "https://github.com/villainy/grpc-accesslog"
documentation = "https://grpc-accesslog.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/villainy/grpc-accesslog/releases"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.56.2"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
bandit = ">=1.7.7"
black = ">=24.1.0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
protobuf = "^4.23.4"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
types-protobuf = "^4.23.0.2"
grpc-stubs = "^1.53.0.5"
pytest-asyncio = "^0.23.6"

[tool.poetry.scripts]
grpc-accesslog = "grpc_accesslog.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = false
source = ["grpc_accesslog", "tests"]
omit = ["tests/proto/*"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [ "grpc" ]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--ignore=tests/proto/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
